Player data loading and saving to Json done. Need to test Generic data saver. need to add functionality for setting data file name

+++ Cyberpunk 2022/Assets/Scripts/IDamageable.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    float Health { get; set;}
+    int Health { get; set;}
     void Damage(int damageAmount);
 }
